plot and stuff

+++ perlmutter/512_nodes_vpicio/srun_server.sh
@@ -1,6 +1,11 @@
 set -x
 set -u
 
+if [ -z "${NUM_NODES}" ]; then
+	echo "Error: NUM_NODES is not set"
+	exit 1
+fi
+
 pushd ./build/bin
 rm server_output.txt || true
 srun -N$NUM_NODES -n$NUM_NODES --ntasks-per-node=1 pkill pdc_server
